Flytt tester som gjeler funkjsoner som ligger i SatsendringUtil til SatsendringUtilTest
